Fix pagination issue
