+++ src/blog/item.njk
@@ -1,4 +1,1 @@
----
-eleventyExcludeFromCollections: true
----
 {{ item.content.rendered | safe }}
