+++ src/now/entry.njk
@@ -1,5 +1,1 @@
----
-eleventyExcludeFromCollections: true
----
-
 {{ entry.content.rendered | safe }}
